# executor: reuse chunk for GetMatchedRowsAndPtrs calls